исправлние работы кода ТТАА и ТТСС, все работает с ипользованием исключений! "ошибка" при декодировании данных кодов не приведт к отказу программы

+++ General_data.h
@@ -71,7 +71,7 @@ public:
 	bool information;                   // Идентификатор существования данных  
 	TEMP_DEWPOINT info_temp;            // Информация о данных температуры и дефекта точки росы
 	WIND wind;                          // Данные по ветру
-	int pressure;                       // Давление на уровне 
+	double pressure;                       // Давление на уровне 
 	surface()
 	{
 		wind_node = false;
